Strategy changed, not debug yet.
