+++ tensorcircuit/gates.py
@@ -6,17 +6,21 @@ import sys
 import warnings
 from copy import deepcopy
 from functools import reduce, partial
-from operator import mul
 from typing import Any, Callable, Optional, Sequence, List, Union, Tuple
+from operator import mul
 
 import numpy as np
+
+try:
+    from numpy import ComplexWarning
+except ImportError:  # np2.0 compatibility
+    from numpy.exceptions import ComplexWarning  # type: ignore
+
 import tensornetwork as tn
 from scipy.stats import unitary_group
 
-from ..cons import backend, dtypestr, npdtype, runtime_backend
-from ..utils import arg_alias
-from . import ComplexWarning
-from . import Gate, num_to_tensor, array_to_tensor
+from .cons import backend, dtypestr, npdtype, runtime_backend
+from .utils import arg_alias
 
 thismodule = sys.modules[__name__]
 
@@ -50,6 +54,7 @@ _wroot_matrix = (
     * np.array([[1, -1 / np.sqrt(2) * (1 + 1.0j)], [1 / np.sqrt(2) * (1 - 1.0j), 1]])
 )
 
+
 _ii_matrix = np.kron(_i_matrix, _i_matrix)
 _xx_matrix = np.kron(_x_matrix, _x_matrix)
 _yy_matrix = np.kron(_y_matrix, _y_matrix)
@@ -68,6 +73,7 @@ _yx_matrix = np.kron(_y_matrix, _x_matrix)
 _yz_matrix = np.kron(_y_matrix, _z_matrix)
 _zx_matrix = np.kron(_z_matrix, _x_matrix)
 _zy_matrix = np.kron(_z_matrix, _y_matrix)
+
 
 _cnot_matrix = np.array(
     [
@@ -105,6 +111,7 @@ _swap_matrix = np.array(
     ]
 )
 
+
 _toffoli_matrix = np.array(
     [
         [1.0, 0, 0, 0, 0, 0, 0, 0],
@@ -140,72 +147,114 @@ def __rmul__(self: tn.Node, lvalue: Union[float, complex]) -> "Gate":
 tn.Node.__rmul__ = __rmul__
 
 
-# def num_to_tensor(*num: Union[float, Tensor], dtype: Optional[str] = None) -> Any:
-#     r"""
-#     Convert the inputs to Tensor with specified dtype.
-#
-#     :Example:
-#
-#     >>> from tensorcircuit.gates import num_to_tensor
-#     >>> # OR
-#     >>> from tensorcircuit.gates import array_to_tensor
-#     >>>
-#     >>> x, y, z = 0, 0.1, np.array([1])
-#     >>>
-#     >>> tc.set_backend('numpy')
-#     numpy_backend
-#     >>> num_to_tensor(x, y, z)
-#     [array(0.+0.j, dtype=complex64), array(0.1+0.j, dtype=complex64), array([1.+0.j], dtype=complex64)]
-#     >>>
-#     >>> tc.set_backend('tensorflow')
-#     tensorflow_backend
-#     >>> num_to_tensor(x, y, z)
-#     [<tf.Tensor: shape=(), dtype=complex64, numpy=0j>,
-#      <tf.Tensor: shape=(), dtype=complex64, numpy=(0.1+0j)>,
-#      <tf.Tensor: shape=(1,), dtype=complex64, numpy=array([1.+0.j], dtype=complex64)>]
-#     >>>
-#     >>> tc.set_backend('pytorch')
-#     pytorch_backend
-#     >>> num_to_tensor(x, y, z)
-#     [tensor(0.+0.j), tensor(0.1000+0.j), tensor([1.+0.j])]
-#     >>>
-#     >>> tc.set_backend('jax')
-#     jax_backend
-#     >>> num_to_tensor(x, y, z)
-#     [DeviceArray(0.+0.j, dtype=complex64),
-#      DeviceArray(0.1+0.j, dtype=complex64),
-#      DeviceArray([1.+0.j], dtype=complex64)]
-#
-#     :param num: inputs
-#     :type num: Union[float, Tensor]
-#     :param dtype: dtype of the output Tensors
-#     :type dtype: str, optional
-#     :return: List of Tensors
-#     :rtype: List[Tensor]
-#     """
-#     # TODO(@YHPeter): fix __doc__ for same function with different names
-#
-#     l = []
-#     if dtype is None:
-#         dtype = dtypestr
-#     for n in num:
-#         if not backend.is_tensor(n):
-#             l.append(backend.cast(backend.convert_to_tensor(n), dtype=dtype))
-#         else:
-#             l.append(backend.cast(n, dtype=dtype))
-#     if len(l) == 1:
-#         return l[0]
-#     return l
-#
-#
-# array_to_tensor = num_to_tensor
+class Gate(tn.Node):  # type: ignore
+    """
+    Wrapper of tn.Node, quantum gate
+    """
+
+    def __repr__(self) -> str:
+        """Formatted output of Gate
+
+        :Example:
+
+        >>> tc.gates.ry(0.5)
+        >>> # OR
+        >>> print(repr(tc.gates.ry(0.5)))
+        Gate(
+            name: '__unnamed_node__',
+            tensor:
+                <tf.Tensor: shape=(2, 2), dtype=complex64, numpy=
+                array([[ 0.9689124 +0.j, -0.24740396+0.j],
+                    [ 0.24740396+0.j,  0.9689124 +0.j]], dtype=complex64)>,
+            edges: [
+                Edge(Dangling Edge)[0],
+                Edge(Dangling Edge)[1]
+            ])
+        """
+        sp = " " * 4
+        edges = self.get_all_edges()
+        edges_text = [edge.__repr__().replace("\n", "").strip() for edge in edges]
+        edges_out = f"[" + f"\n{sp*2}" + f",\n{sp*2}".join(edges_text) + f"\n{sp}]"
+        tensor_out = f"\n{sp*2}" + self.tensor.__repr__().replace("\n", f"\n{sp*2}")
+        return (
+            f"{self.__class__.__name__}(\n"
+            f"{sp}name: {self.name!r},\n"
+            f"{sp}tensor:{tensor_out},\n"
+            f"{sp}edges: {edges_out})"
+        )
+
+    def copy(self, conjugate: bool = False) -> "Gate":
+        result = super().copy(conjugate=conjugate)
+        result.__class__ = Gate
+        return result  # type: ignore
 
 
-# def gate_wrapper(m: Tensor, n: Optional[str] = None) -> Gate:
-#     if n is None:
-#         n = "unknowngate"
-#     m = m.astype(npdtype)
-#     return Gate(deepcopy(m), name=n)
+def num_to_tensor(*num: Union[float, Tensor], dtype: Optional[str] = None) -> Any:
+    r"""
+    Convert the inputs to Tensor with specified dtype.
+
+    :Example:
+
+    >>> from tensorcircuit.gates import num_to_tensor
+    >>> # OR
+    >>> from tensorcircuit.gates import array_to_tensor
+    >>>
+    >>> x, y, z = 0, 0.1, np.array([1])
+    >>>
+    >>> tc.set_backend('numpy')
+    numpy_backend
+    >>> num_to_tensor(x, y, z)
+    [array(0.+0.j, dtype=complex64), array(0.1+0.j, dtype=complex64), array([1.+0.j], dtype=complex64)]
+    >>>
+    >>> tc.set_backend('tensorflow')
+    tensorflow_backend
+    >>> num_to_tensor(x, y, z)
+    [<tf.Tensor: shape=(), dtype=complex64, numpy=0j>,
+     <tf.Tensor: shape=(), dtype=complex64, numpy=(0.1+0j)>,
+     <tf.Tensor: shape=(1,), dtype=complex64, numpy=array([1.+0.j], dtype=complex64)>]
+    >>>
+    >>> tc.set_backend('pytorch')
+    pytorch_backend
+    >>> num_to_tensor(x, y, z)
+    [tensor(0.+0.j), tensor(0.1000+0.j), tensor([1.+0.j])]
+    >>>
+    >>> tc.set_backend('jax')
+    jax_backend
+    >>> num_to_tensor(x, y, z)
+    [DeviceArray(0.+0.j, dtype=complex64),
+     DeviceArray(0.1+0.j, dtype=complex64),
+     DeviceArray([1.+0.j], dtype=complex64)]
+
+    :param num: inputs
+    :type num: Union[float, Tensor]
+    :param dtype: dtype of the output Tensors
+    :type dtype: str, optional
+    :return: List of Tensors
+    :rtype: List[Tensor]
+    """
+    # TODO(@YHPeter): fix __doc__ for same function with different names
+
+    l = []
+    if dtype is None:
+        dtype = dtypestr
+    for n in num:
+        if not backend.is_tensor(n):
+            l.append(backend.cast(backend.convert_to_tensor(n), dtype=dtype))
+        else:
+            l.append(backend.cast(n, dtype=dtype))
+    if len(l) == 1:
+        return l[0]
+    return l
+
+
+array_to_tensor = num_to_tensor
+
+
+def gate_wrapper(m: Tensor, n: Optional[str] = None) -> Gate:
+    if n is None:
+        n = "unknowngate"
+    m = m.astype(npdtype)
+    return Gate(deepcopy(m), name=n)
 
 
 class GateF:
@@ -361,66 +410,66 @@ meta_gate()
 pauli_gates = [i(), x(), y(), z()]  # type: ignore
 
 
-# def matrix_for_gate(gate: Gate, tol: float = 1e-6) -> Tensor:
-#     r"""
-#     Convert Gate to numpy array.
-#
-#     :Example:
-#
-#     >>> gate = tc.gates.r_gate()
-#     >>> tc.gates.matrix_for_gate(gate)
-#         array([[1.+0.j, 0.+0.j],
-#             [0.+0.j, 1.+0.j]], dtype=complex64)
-#
-#     :param gate: input Gate
-#     :type gate: Gate
-#     :return: Corresponding Tensor
-#     :rtype: Tensor
-#     """
-#
-#     t = gate.tensor
-#     t = backend.reshapem(t)
-#     t = backend.numpy(t)
-#     t.real[abs(t.real) < tol] = 0.0
-#     t.imag[abs(t.imag) < tol] = 0.0
-#     return t
+def matrix_for_gate(gate: Gate, tol: float = 1e-6) -> Tensor:
+    r"""
+    Convert Gate to numpy array.
+
+    :Example:
+
+    >>> gate = tc.gates.r_gate()
+    >>> tc.gates.matrix_for_gate(gate)
+        array([[1.+0.j, 0.+0.j],
+            [0.+0.j, 1.+0.j]], dtype=complex64)
+
+    :param gate: input Gate
+    :type gate: Gate
+    :return: Corresponding Tensor
+    :rtype: Tensor
+    """
+
+    t = gate.tensor
+    t = backend.reshapem(t)
+    t = backend.numpy(t)
+    t.real[abs(t.real) < tol] = 0.0
+    t.imag[abs(t.imag) < tol] = 0.0
+    return t
 
 
-# def bmatrix(a: Array) -> str:
-#     r"""
-#     Returns a :math:`\LaTeX` bmatrix.
-#
-#     :Example:
-#
-#     >>> gate = tc.gates.r_gate()
-#     >>> array = tc.gates.matrix_for_gate(gate)
-#     >>> array
-#     array([[1.+0.j, 0.+0.j],
-#         [0.+0.j, 1.+0.j]], dtype=complex64)
-#     >>> print(tc.gates.bmatrix(array))
-#     \begin{bmatrix}    1.+0.j & 0.+0.j\\    0.+0.j & 1.+0.j \end{bmatrix}
-#
-#     Formatted Display:
-#
-#     .. math::
-#         \begin{bmatrix}    1.+0.j & 0.+0.j\\    0.+0.j & 1.+0.j \end{bmatrix}
-#
-#     :param a: 2D numpy array
-#     :type a: np.array
-#     :raises ValueError: ValueError("bmatrix can at most display two dimensions")
-#     :return: :math:`\LaTeX`-formatted string for bmatrix of the array a
-#     :rtype: str
-#     """
-#     #   Adopted from https://stackoverflow.com/questions/17129290/numpy-2d-and-1d-array-to-latex-bmatrix/17131750
-#
-#     if len(a.shape) > 2:
-#         raise ValueError("bmatrix can at most display two dimensions")
-#     lines = str(a).replace("[", "").replace("]", "").splitlines()
-#     rv = [r"\begin{bmatrix}"]
-#     rv += ["    " + " & ".join(l.split()) + r"\\" for l in lines]
-#     rv[-1] = rv[-1][:-2]
-#     rv += [r" \end{bmatrix}"]
-#     return "".join(rv)
+def bmatrix(a: Array) -> str:
+    r"""
+    Returns a :math:`\LaTeX` bmatrix.
+
+    :Example:
+
+    >>> gate = tc.gates.r_gate()
+    >>> array = tc.gates.matrix_for_gate(gate)
+    >>> array
+    array([[1.+0.j, 0.+0.j],
+        [0.+0.j, 1.+0.j]], dtype=complex64)
+    >>> print(tc.gates.bmatrix(array))
+    \begin{bmatrix}    1.+0.j & 0.+0.j\\    0.+0.j & 1.+0.j \end{bmatrix}
+
+    Formatted Display:
+
+    .. math::
+        \begin{bmatrix}    1.+0.j & 0.+0.j\\    0.+0.j & 1.+0.j \end{bmatrix}
+
+    :param a: 2D numpy array
+    :type a: np.array
+    :raises ValueError: ValueError("bmatrix can at most display two dimensions")
+    :return: :math:`\LaTeX`-formatted string for bmatrix of the array a
+    :rtype: str
+    """
+    #   Adopted from https://stackoverflow.com/questions/17129290/numpy-2d-and-1d-array-to-latex-bmatrix/17131750
+
+    if len(a.shape) > 2:
+        raise ValueError("bmatrix can at most display two dimensions")
+    lines = str(a).replace("[", "").replace("]", "").splitlines()
+    rv = [r"\begin{bmatrix}"]
+    rv += ["    " + " & ".join(l.split()) + r"\\" for l in lines]
+    rv[-1] = rv[-1][:-2]
+    rv += [r" \end{bmatrix}"]
+    return "".join(rv)
 
 
 def phase_gate(theta: float = 0) -> Gate:
@@ -727,7 +776,7 @@ def random_two_qubit_gate() -> Gate:
     return Gate(deepcopy(unitary), name="R2Q")
 
 
-def any_gate(unitary: Tensor, name: str = "any") -> Gate:
+def any_gate(unitary: Tensor, name: str = "any", dim: Optional[int] = None) -> Gate:
     """
     Note one should provide the gate with properly reshaped.
 
@@ -735,6 +784,8 @@ def any_gate(unitary: Tensor, name: str = "any") -> Gate:
     :type unitary: Tensor
     :param name: The name of the gate.
     :type name: str
+    :param dim: The dimension of the gate.
+    :type dim: int
     :return: the resulted gate
     :rtype: Gate
     """
@@ -743,7 +794,10 @@ def any_gate(unitary: Tensor, name: str = "any") -> Gate:
         unitary.tensor = backend.cast(unitary.tensor, dtypestr)
         return unitary
     unitary = backend.cast(unitary, dtypestr)
-    unitary = backend.reshape2(unitary)
+    if dim is None:
+        unitary = backend.reshape2(unitary)
+    else:
+        unitary = backend.reshaped(unitary, dim)
     # nleg = int(np.log2(backend.sizen(unitary)))
     # unitary = backend.reshape(unitary, [2 for _ in range(nleg)])
     return Gate(unitary, name=name)
@@ -777,8 +831,6 @@ def exponential_gate(unitary: Tensor, theta: float, name: str = "none") -> Gate:
 
 
 exp_gate = exponential_gate
-
-
 # exp = exponential_gate
 
 
@@ -887,7 +939,7 @@ def multicontrol_gate(unitary: Tensor, ctrl: Union[int, Sequence[int]] = 1) -> O
         for i in range(1, len(nodes) - 1):
             nodes[i][3] ^ nodes[i + 1][0]
 
-    from ..quantum import QuOperator
+    from .quantum import QuOperator
 
     l = int((len(nodes[-1].edges) - 1) / 2)
     gate = QuOperator(
